Updated Login and Registration Page (UI/UX)
